+++ mskpy/ephem.py
@@ -1,0 +1,34 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+ephem --- Ephemeris tools
+=========================
+
+Requres PySPICE.
+
+.. autosummary::
+   :toctree: generated/
+
+   MovingTarget
+
+   getxyz
+
+"""
+
+__all__ = [
+    ''
+]
+
+import numpy as np
+import astropy.units as u
+from astropy.units import Quantity
+import spice as pyspice
+
+class MovingObject():
+    pass
+
+
+def getxyz()
+    """Coordinates and velocity from an ephemeris kernel.
+    """
+    pass
+